+++ changelog.md
@@ -4,6 +4,7 @@
 
 - Add authorization in GitHub API usage to avoid API limits
 - Using custom protobuf installation
+- Add `VersionCheck` and `ForcePath` lib `Error` variants
 
 Thanks: